Added vcs=2 to some of the API calls
A recent update to Kiln allows you to choose the VCS of a repository on creation, i.e., Mercurial only,
Git only, or Harmony (Git & Mercurial).

Since this update, it is necessary to add &vcs=2 to the URL for Git-only repositories. See, for example,
the documentation for Api/{version}/Review/{sReview} (GET) here...
https://developers.fogbugz.com/?W209

NOTE: specifying 'vcs=2' will work for Harmony repositories and Git-only repositories, but it won't
work for Mercurial-only repositories.